Allow to set the game to be run via build arguments
Currently, the Dockerfile really only supports to run street fighter. The
necessary changes to run a different game are however rather small and can be
set via build arguments instead of hardcoding them.
Also, don't rely on a patched docker and instead provide the `SCCcredentials`
file as a secret to the container build.